feat(stage3e2de): refacto du candidaterprops fixture en donneesEntreprise
